Persist elasticsearch indexes and safer startup for elasticsearch and kibana

+++ develop.sh
@@ -51,10 +51,10 @@ else
   echo "You have no config.sh. Copy config.defaults.sh to config.sh, change the passwords and try again." && exit 1;
 fi
 
-docker stop elasticsearch
-docker rm elasticsearch
-docker stop kibana
-docker rm kibana
+[ "$(docker ps | grep elasticsearch)" ] && docker stop elasticsearch 
+[ "$(docker ps -a | grep elasticsearch)" ] && docker rm elasticsearch 
+[ "$(docker ps | grep kibana)" ] && docker stop kibana 
+[ "$(docker ps -a | grep kibana)" ] && docker rm kibana 
 
 docker run -d \
   -v $(pwd)/elasticsearch/config/elasticsearch.yml:/usr/share/elasticsearch/config/elasticsearch.yml:ro \
@@ -62,6 +62,7 @@ docker run -d \
   -p 9300:9300 \
   -e ES_JAVA_OPTS="-Xmx256m -Xms256m" \
   --name elasticsearch \
+  --volume $(pwd)/data/elasticsearch:/usr/share/elasticsearch/data \
   docker.elastic.co/elasticsearch/elasticsearch-oss:6.3.0
 
 docker run -d \
